App previews with built in NFT
