+++ functions/[[accountId]]/widget/[[index]].js
@@ -1,3 +1,5 @@
+import { Buffer } from "node:buffer";
+
 class MetaTitleInjector {
   constructor({ title }) {
     this.title = title;
@@ -9,13 +11,16 @@ class MetaTitleInjector {
 }
 
 class MetaImageInjector {
-  constructor({ image }) {
+  constructor({ image, authorImage }) {
     this.image = image;
+    this.authorImage = authorImage;
   }
 
   element(element) {
     if (this.image) {
       element.setAttribute("content", this.image);
+    } else if (this.authorImage) {
+      element.setAttribute("content", this.authorImage);
     }
   }
 }
@@ -106,11 +111,84 @@ async function socialGet(keys, blockHeight, parse) {
   return data;
 }
 
+async function viewCall({ contractId, method, args }) {
+  const res = await fetch("https://rpc.mainnet.near.org", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({
+      jsonrpc: "2.0",
+      id: "dontcare",
+      method: "query",
+      params: {
+        request_type: "call_function",
+        finality: "final",
+        account_id: contractId,
+        method_name: method,
+        args_base64: btoa(JSON.stringify(args)),
+      },
+    }),
+  });
+  const json = await res.json();
+  const result = Buffer.from(json.result.result).toString("utf-8");
+  return JSON.parse(result);
+}
+
+async function nftToImageUrl({ contractId, tokenId }) {
+  const [token, nftMetadata] = await Promise.all([
+    viewCall({
+      contractId,
+      method: "nft_token",
+      args: { token_id: tokenId },
+    }),
+    viewCall({
+      contractId,
+      method: "nft_metadata",
+      args: {},
+    }),
+  ]);
+
+  const tokenMetadata = token?.metadata || {};
+  const tokenMedia = tokenMetadata.media || "";
+
+  let imageUrl =
+    tokenMedia.startsWith("https://") ||
+    tokenMedia.startsWith("http://") ||
+    tokenMedia.startsWith("data:image")
+      ? tokenMedia
+      : nftMetadata.base_uri
+      ? `${nftMetadata.base_uri}/${tokenMedia}`
+      : tokenMedia.startsWith("Qm") || tokenMedia.startsWith("ba")
+      ? `https://ipfs.near.social/ipfs/${tokenMedia}`
+      : tokenMedia;
+
+  if (!tokenMedia && tokenMetadata.reference) {
+    const metadataUrl =
+      nftMetadata.base_uri === "https://arweave.net" &&
+      !tokenMetadata.reference.startsWith("https://")
+        ? `${nftMetadata.base_uri}/${tokenMetadata.reference}`
+        : tokenMetadata.reference.startsWith("https://") ||
+          tokenMetadata.reference.startsWith("http://")
+        ? tokenMetadata.reference
+        : tokenMetadata.reference.startsWith("ar://")
+        ? `https://arweave.net/${tokenMetadata.reference.split("//")[1]}`
+        : null;
+    if (metadataUrl) {
+      const res = await fetch(metadataUrl);
+      const json = await res.json();
+      imageUrl = json.media;
+    }
+  }
+
+  return imageUrl;
+}
+
 function wrapImage(url) {
   return url ? `https://i.near.social/large/${url}` : null;
 }
 
-async function internalImageToUrl(image) {
+async function internalImageToUrl(env, image) {
   if (image?.url) {
     return image.url;
   } else if (image?.ipfs_cid) {
@@ -118,22 +196,28 @@ async function internalImageToUrl(image) {
   } else if (image?.nft) {
     try {
       const { contractId, tokenId } = image.nft;
-      const request = await fetch(
-        `https://near.social/magic/nft/${contractId}/${tokenId}`
-      );
-      return request.text();
+      const NftKV = env.NftKV;
+
+      let imageUrl = await NftKV.get(`${contractId}/${tokenId}`);
+      if (!imageUrl) {
+        imageUrl = await nftToImageUrl({ contractId, tokenId });
+        if (imageUrl) {
+          await NftKV.put(`${contractId}/${tokenId}`, imageUrl);
+        }
+      }
+      return imageUrl;
     } catch (e) {
-      return null;
+      console.log(e);
     }
   }
   return null;
 }
 
-async function imageToUrl(image) {
-  return wrapImage(await internalImageToUrl(image));
+async function imageToUrl(env, image) {
+  return wrapImage(await internalImageToUrl(env, image));
 }
 
-async function postData(url, parts, data, isPost) {
+async function postData(env, url, data, isPost) {
   const accountId = url.searchParams.get("accountId");
   const blockHeight = url.searchParams.get("blockHeight");
   const [content, name, authorImage] = await Promise.all([
@@ -148,9 +232,9 @@ async function postData(url, parts, data, isPost) {
 
   data.raw = content;
   data.description = content?.text || "";
-  data.image = await imageToUrl(content?.image);
+  data.image = await imageToUrl(env, content?.image);
   if (!data.image) {
-    data.authorImage = await imageToUrl(authorImage);
+    data.authorImage = await imageToUrl(env, authorImage);
   }
   data.title = isPost
     ? `Post by ${name ?? accountId} | Near Social`
@@ -159,15 +243,19 @@ async function postData(url, parts, data, isPost) {
   data.accountId = accountId;
 }
 
-async function profileData(url, parts, data) {
+async function profileData(env, url, data) {
   const accountId = url.searchParams.get("accountId");
   const profile = await socialGet(`${accountId}/profile/**`);
 
   const name = profile?.name;
   data.raw = profile;
   data.description = profile?.description || "";
-  data.image = await imageToUrl(profile?.image);
-  data.authorImage = data.image;
+  data.image = await imageToUrl(env, profile?.image);
+  data.authorImage =
+    data.image ||
+    wrapImage(
+      "https://ipfs.near.social/ipfs/bafkreibmiy4ozblcgv3fm3gc6q62s55em33vconbavfd2ekkuliznaq3zm"
+    );
   data.title = name
     ? `${name} (${accountId}) | Near Social`
     : `${accountId} | Near Social`;
@@ -175,16 +263,34 @@ async function profileData(url, parts, data) {
   data.accountId = accountId;
 }
 
-async function generateData(url) {
-  const data = defaultData();
+async function widgetData(env, url, data) {
   const parts = url.pathname.split("/");
+  const accountId = parts[1];
+  const widgetId = parts[3];
+  const metadata = await socialGet(
+    `${accountId}/widget/${widgetId}/metadata/**`
+  );
+
+  const name = metadata?.name || widgetId;
+  data.raw = metadata;
+  data.description = metadata?.description || "";
+  data.image = await imageToUrl(env, metadata?.image);
+  data.title = `${name} by ${accountId} | Near Social`;
+  data.accountName = name;
+  data.accountId = accountId;
+}
+
+async function generateData(env, url) {
+  const data = defaultData();
   try {
     if (url.pathname === "/mob.near/widget/MainPage.Post.Page") {
-      await postData(url, parts, data, true);
+      await postData(env, url, data, true);
     } else if (url.pathname === "/mob.near/widget/MainPage.Comment.Page") {
-      await postData(url, parts, data, false);
+      await postData(env, url, data, false);
     } else if (url.pathname === "/mob.near/widget/ProfilePage") {
-      await profileData(url, parts, data);
+      await profileData(env, url, data);
+    } else {
+      await widgetData(env, url, data);
     }
   } catch (e) {
     console.error(e);
@@ -194,12 +300,12 @@ async function generateData(url) {
   return data;
 }
 
-export async function onRequest({ request, next }) {
+export async function onRequest({ request, next, env }) {
   const url = new URL(request.url);
   if (url.pathname.split("/").length < 4) {
     return next();
   }
-  const data = await generateData(url);
+  const data = await generateData(env, url);
   return (
     new HTMLRewriter()
       .on('meta[property="og:title"]', new MetaTitleInjector(data))
